Initial structure for adding levels to stories.
Please don't take this as a standar for future development
until it get merged in develop branch.

This branch its pretty edgy.

The development of this structure is in progress.

This files need formatting but as I'm not going to work on it
during this weekend I'm pushing them in order to see the
changes and get the general idea by the rest of the team.